lib/brillo/adapter/postgres: Respect `database.url` configuration.
The intention is to still work exactly the same as before if users are using `user`, `password`,
etc. from `database.yml`, but now also support the `url` field which if i understand correctly
should take precedence.

I've attempted to write this in such a way that existing behaviour is preserved.  In particular, the
log anonimisation technique is handy, so i'm pulling the password out of the URI.